Add YAML for deploying VecTextSearch and Weaviate on Kubernetes
This commit adds example YAML files for deploying VecTextSearch and Weaviate on a Kubernetes cluster. The VecTextSearch YAML includes a Deployment, Service, and Ingress, while the Weaviate YAML includes only a Deployment and Service. These files can be used as a starting point for deploying the VecTextSearch and Weaviate applications to Kubernetes.

Author: ChatGPT